added profile image display at base
The base.html is copied to index.html of user profile app and is edited
to include profile image, which displays a placeholder image if no
profile image is located. Also added a profilepage.html to prepare to
make a page that display the profile info for the users. Note: urls.py
in search app is edited but not included in this commit